+++ maeumgagym-infrastructure/src/test/kotlin/com/info/maeumgagym/domain/quote/QuoteServiceTests.kt
@@ -1,4 +1,4 @@
-package com.info.maeumgagym.domain.quote.junit5
+package com.info.maeumgagym.domain.quote
 
 import com.info.maeumgagym.domain.quote.exception.MismatchQuoteAndQuoterException
 import com.info.maeumgagym.quote.port.`in`.ReadRandomQuoteUseCase
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional
 
 @Transactional
 @SpringBootTest
-class QuoteTests @Autowired constructor(
+class QuoteServiceTests @Autowired constructor(
     private val readRandomQuoteUseCase: ReadRandomQuoteUseCase
 ) {
