+++ maeumgagym-infrastructure/src/test/kotlin/com/info/maeumgagym/domain/routine/UpdateRoutineServiceTests.kt
@@ -1,19 +1,17 @@
-package com.info.maeumgagym.domain.routine.junit5
+package com.info.maeumgagym.domain.routine
 
 import com.info.maeumgagym.auth.exception.PermissionDeniedException
 import com.info.maeumgagym.domain.auth.AuthTestModule.saveInContext
 import com.info.maeumgagym.domain.routine.entity.RoutineJpaEntity
-import com.info.maeumgagym.domain.routine.module.RoutineTestModule
-import com.info.maeumgagym.domain.routine.module.RoutineTestModule.saveInRepository
+import com.info.maeumgagym.domain.routine.RoutineTestModule.saveInRepository
 import com.info.maeumgagym.domain.routine.repository.RoutineRepository
 import com.info.maeumgagym.domain.user.entity.UserJpaEntity
 import com.info.maeumgagym.domain.user.mapper.UserMapper
-import com.info.maeumgagym.domain.user.module.UserTestModule
-import com.info.maeumgagym.domain.user.module.UserTestModule.saveInContext
-import com.info.maeumgagym.domain.user.module.UserTestModule.saveInRepository
+import com.info.maeumgagym.domain.user.UserTestModule
+import com.info.maeumgagym.domain.user.UserTestModule.saveInRepository
 import com.info.maeumgagym.domain.user.repository.UserRepository
 import com.info.maeumgagym.routine.exception.RoutineNotFoundException
-import com.info.maeumgagym.routine.port.`in`.DeleteRoutineUseCase
+import com.info.maeumgagym.routine.port.`in`.UpdateRoutineUseCase
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -21,48 +19,54 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.transaction.annotation.Transactional
+import javax.persistence.EntityManager
 
 @Transactional
 @SpringBootTest
-class DeleteRoutineServiceTests @Autowired constructor(
-    private val deleteRoutineUseCase: DeleteRoutineUseCase,
+class UpdateRoutineServiceTests @Autowired constructor(
+    private val updateRoutineUseCase: UpdateRoutineUseCase,
     private val routineRepository: RoutineRepository,
     private val userRepository: UserRepository,
-    private val userMapper: UserMapper
+    private val userMapper: UserMapper,
+    private val entityManager: EntityManager
 ) {
+
     private lateinit var user: UserJpaEntity
-    private lateinit var otherUser: UserJpaEntity
     private lateinit var routine: RoutineJpaEntity
 
     @BeforeEach
     fun initialize() {
         user = UserTestModule.createTestUser().saveInRepository(userRepository).saveInContext(userMapper)
-        otherUser = UserTestModule.createOtherUser().saveInRepository(userRepository)
         routine = RoutineTestModule.createTestRoutine(user.id!!).saveInRepository(routineRepository)
     }
 
     @Test
-    fun deleteMyRoutine() {
-        Assertions.assertDoesNotThrow {
-            deleteRoutineUseCase.deleteRoutine(routine.id!!)
-        }
-        Assertions.assertNull(routineRepository.findByIdOrNull(routine.id!!))
+    fun updateMyRoutine() {
+        entityManager.detach(routine)
+        val request = RoutineTestModule.getUpdateRoutineRequest(routine)
+        updateRoutineUseCase.updateRoutine(request, routine.id!!)
+        Assertions.assertNotEquals(routine, routineRepository.findByIdOrNull(routine.id!!))
     }
 
     @Test
-    fun deleteOtherRoutine() {
-        otherUser.saveInContext(userMapper)
+    fun updateOtherRoutine() {
+        UserTestModule.createOtherUser().saveInRepository(userRepository).saveInContext(userMapper)
         Assertions.assertThrows(PermissionDeniedException::class.java) {
-            deleteRoutineUseCase.deleteRoutine(routine.id!!)
+            updateRoutineUseCase.updateRoutine(
+                RoutineTestModule.getUpdateRoutineRequest(routine),
+                routine.id!!
+            )
         }
-        Assertions.assertNotNull(routineRepository.findByIdOrNull(routine.id!!))
     }
 
     @Test
-    fun deleteNonExistentRoutine() {
+    fun updateNonExistentRoutine() {
         routineRepository.deleteById(routine.id!!)
         Assertions.assertThrows(RoutineNotFoundException::class.java) {
-            deleteRoutineUseCase.deleteRoutine(routine.id!!)
+            updateRoutineUseCase.updateRoutine(
+                RoutineTestModule.getUpdateRoutineRequest(routine),
+                routine.id!!
+            )
         }
     }
 }
